Fixed a bug inside the PlayerState machine
The problem was the following: I extracted some methods for e.g. fallDown left and right to
the parent class. But that meant, if the player was in fallDownLeft state, the method was
called and a state transition from fallDownLeft to fallDownLeft was executed (but not
necessary). This problem I found during implementing double jump. So I fixed it. now every
state has it's named transitions to follow-up states.

Unfortunately double jump doesn't work at the moment, because I ran into another problem...
An import from the user/player can take multiple game loop cycles (e.g. make a jump)
which is a problem by counting the jumps (a cycle with vsync on takes about 0.017 sec, with
vsync of it's 0.001 sec - so I need a solution how to check when a player input is done,
somehow)